+++ src/index.ts
@@ -6,5 +6,5 @@ export { default as Intercom } from './components/Intercom'
 export { default as Messenger } from './components/Messenger'
 export { default as Userlike } from './components/Userlike'
 export { default as Chatwoot } from './components/Chatwoot'
-export { default as FrontChat } from './components/FrontChat'
+export { default as Front } from './components/Front'
 export { default as LiveChatLoaderProvider } from './components/LiveChatLoaderProvider'
