+++ src/providers/index.ts
@@ -4,4 +4,4 @@ export { default as intercom } from './intercom'
 export { default as messenger } from './messenger'
 export { default as userlike } from './userlike'
 export { default as chatwoot } from './chatwoot'
-export { default as frontChat } from './frontChat'
+export { default as front } from './front'
