+++ website/pages/front.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { LiveChatLoaderProvider, FrontChat } from 'react-live-chat-loader'
+import { LiveChatLoaderProvider, Front } from 'react-live-chat-loader'
 import type { NextPage } from 'next'
 
 import Layout from '../layouts/main'
@@ -7,15 +7,15 @@ import ExampleLinks from '../components/exampleLinks'
 
 const Page: NextPage = () => (
   <LiveChatLoaderProvider
-    provider="frontChat"
-    providerKey="68cdc08b45c2d546d827398bb8d0d857"
+    provider="front"
+    providerKey="b2b8dcd0c63c934d363e6704cfa6af86"
   >
-    <Layout title="React Live Chat Loader: FrontChat">
+    <Layout title="React Live Chat Loader: Front">
       <div className="wrapper">
         <div className="inner">
-          <h1>React Live Chat Loader: FrontChat</h1>
+          <h1>React Live Chat Loader: Front</h1>
           <p>
-            This is an example implementation of the FrontChat chat widget using{' '}
+            This is an example implementation of the Front chat widget using{' '}
             <a href="https://github.com/calibreapp/react-live-chat-loader">
               react-live-chat-loader
             </a>
@@ -25,7 +25,7 @@ const Page: NextPage = () => (
           <ExampleLinks />
         </div>
       </div>
-      <FrontChat color="#333333" />
+      <Front color="#333333" />
     </Layout>
   </LiveChatLoaderProvider>
 )
